attack -> 확률값 계산을 실수형으로 변경하여 스킬의 확률을 좀 더 낮추고 평타의 확률을 높였다. move -> 테스트 변경 battleRoar -> 모션이 끝나면 다시 move패턴으로 이동하고, 랜덤하게 지역 광역기 / 스턴기를 사용하여 도트데미지, 스턴을 넣도록 기능 구현

+++ 팀 프레임워크/bossActionAttack.h
@@ -7,7 +7,7 @@ private:
 	float passedTime;
 	float activeTime;
 
-	int resultValue;
+	float resultValue;
 	vector<int> list;
 public:
 	bossActionAttack();
